+++ site/jaws_site/models.py
@@ -47,6 +47,8 @@ class Run(Base):
     )
     workflow_root = Column(String(1024), nullable=True)
     workflow_name = Column(String(64), nullable=True)
+    wdl_basename = Column(String(64), nullable=False)
+    json_basename = Column(String(64), nullable=False)
 
 
 class Run_Log(Base):