+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - ensure infiles exist upon (re)submit before sending to Cromwell; i.e. haven't been purged (jaws-support#110)
 - update atime for infiles when (re)submit; to avoid purging files prematurely (#1689)
 - outputs.json shall contain relpaths instead of abspaths (#1652)
+- runs' output folders shall once again contain a copy of the WDL, inputs-JSON, and subworkflows-ZIP (#1710)
 
 ## 2.0.0 Summary [08/23/2023]
 ### Internal Facing Changes